Overhaul of "Behavior not considered unsafe"
Give this section of the reference a new paint job :3

Note: I’m not entirely sure if this section should even exist permanently. It doesn’t really fit in with strictly defining Rust’s behavior as such, since it describes lots of legal behavior that is simply surprising to users. It may fit better in the nomicon or a later part of the book. On the other hand, neither of those would suffer from the improvements made here.

- Slightly rename the section to clarify that we refer to generally undesirable behavior.
- Adopt language to authoritative spec language (do not use we, do not refer to implementations, use as specific of a language as possible)
- Link to standard library
- Link to applicable other sections
- Use admonitions for existing text where applicable
- Insert rule identifiers everywhere
- Write a complete section for deadlocks, including a deadlocking code example
- Expand the logic bug example with Hash and Eq into a full code example