WIP: Revise editorially
Let's make some editorial revisions to clarify wording and add and
clarify examples.

One thing we're struggling with a bit here is how much we want to
overlap with the documentation of the standard library.  In some
cases, we've removed detailed descriptions of the implementation rule
when this is better covered in the library docs.

Thanks to Eric Huss for the glossary entry added here.