Revise `proc_macro_derive` text
One thing worth highlighting is that we're preferring to say that we
"declare" (rather than "define") derive macro helper attributes.  A
definition is something that says what something is.  For these
attributes, though, we're just declaring that they exist.

In revising, we've clarified that the derive macro function must
appear in the root of the crate and that it may be `const` and may use
`extern` (with the allowed ABI) but may not use other qualifiers.

We've also changed the syntax to use `IDENTIFIER` rather than
`SimplePathSegment`.  In testing, none of the other things that would
be accepted by the latter are allowed, so it seems more correct in
terms of user-visible behavior to use `IDENTIFIER` here.