RISC-V: Fix (rename) `Zb` to `B`
The `Zb` extension does not exist and we instead have the `B` extension
which is a superset of the three subextensions: `Zba`, `Zbb` and `Zbs`.

This commit fixes this issue and updates the reference URL to the source
code of the latest ratified ISA Manual (version 20250508).

To align with rust-lang/rust#145948,
this commit performs rename, not removal.