Revise and extend text on extending expressions
In particular, let's note that the final expression of an async block
expression is not extending; we need to say that as we otherwise say
that the final expression of a block expression is extending, and
async blocks are ostensibly blocks in our ontology.

Let's build the pass tests with Rust 2024 to ensure that we're not
relying on the old edition tail expression rule.

When talking about an arm of an extending match expression, let's be
clear that we're talking about the arm expression.

Let's break apart all of the failure tests; otherwise we can't be sure
that they're all failing, and let's mark the expected error code for
each failure.

For each test, let's note the applicable rule for the outermost part
of the expression.

Let's add some tests to show that the operand of `break` is not
extending, as that might be a bit surprising.

Let's demonstrate explicitly that raw borrows count as borrows.