Enable folding of chapter listing in navigation sidebar
We have many chapters and subchapters.  Listing all of the subchapters
in the navigation sidebar by default can make it difficult to see the
high level structure of the Reference and can inhibit rapid navigation
based on that hierarchical structure.

Fortunately, mdbook has an option to fold the chapter listing.  This
causes only the top-level chapters to be shown by default while giving
the user the ability to unfold individual chapters.  Let's turn on
that option.