Link out to the notation from grammar summary
As a reference, users are likely to jump in at random points and it can
then be hard to know where to look for information on grammars.

While it would likely be overkill to link to the notation wherever there
is a grammar, the grammary summary page seems like a reasonable place to
at least include a link.