Add rule identifiers to notation chapter
We originally hadn't added rule identifiers to the notation chapter as
in some sense these aren't language rules.  However, as it turns out,
it's still useful to have these so as to be able to refer by name to
particular sections.  And anyway, since the language rules are defined
in terms of this notation, the notation rules are in a sense normative
at a meta level.

In doing this, we'll nest certain sections under the grammar section
that should be so nested.