Move `input.syntax` to the corresponding chapter
We had a set of grammar rules we had placed in the notation chapter
under the label of "common productions".  There's a certain sense to
that, particularly if we had more of them.  As it is, though, all of
the productions seem related enough to the "input format" chapter, and
in fact we had already labeled these productions with the identifier
`input.syntax`, so let's go ahead and move these productions to the
"input format" chapter.  This looks more orderly, especially as we now
have rule labels in the notation chapter.