Add support to grammar for single line comments
Sometimes it would be useful in the grammar to be able to add comments
inline.  Let's allow this.

We'll treat these comments as expressions and add special handling to
allow these on lines ahead of the production name.

For now we'll render these only in the Markdown and not in the
railroad diagrams.