Specify the target limits for target-specific ABIs
This specifies the restrictions on which target-specific ABIs are
available. These are defined in `rustc` in [`is_abi_supported`] and
the associated `AbiMap`.

Some of these are currently only implemented as future incompatibly
warnings via `unsupported_calling_conventions`. I'm fine with just not
acknowledging that and saying what the intent is.

This is split off from https://github.com/rust-lang/reference/pull/1948
with the intent of separating out the changes.

[`is_abi_supported`]: https://github.com/rust-lang/rust/blob/eec6bd9d69832f57341c6de6a93fa7b9f47e2111/compiler/rustc_target/src/spec/mod.rs#L2610-L2613

Edited-by: TC