More clearly define the default abi of an extern block
This attempts to make it clearer that if the ABI string is not
specified, it defaults to the "C" ABI. I felt like the existing text
didn't emphasize that enough.

I also took the opportunity to reword the intro to more generally
explain what the ABI string does.

A followup PR will further clarify some of the confusing use of the word
"default" in the specific ABI definitions.