Combine the two const function sections
There were two sections defining const functions, which was a little
confusing to me and made it a little difficult to maintain them. This
combines them so that there is only one section. It could have gone
in either chapter, but I arbitrarily picked const-eval.

This removes `items.fn.const.extern`. The reference generally assumes
that if something can be defined, it can be used unless explicitly
restricted. This used to be a restriction, but it morphed over time as
`extern` support was slowly added.

I noticed that `const-eval.const-fn.type-restrictions` is not clear to
me and needs further work to explain the type restrictions.