Revise note about `{Path,Identifier}Pattern` parsing
Let's extend the example to demonstrate, comparatively, what is parsed
as an `IdentifierPattern`.  Even though this is the common case,
showing this might be helpful to people less familiar with the
grammar.  Let's remove `main`, as that's not needed here.  And, in the
code, let's highlight specifically what we're commenting on by
commenting below the construct.