Clarify default when not using `extern fn`
Due to changes we've recently merged discussing the default ABI of an
`extern` that doesn't specify one ("C"), the description of "Rust"
being the default ABI for functions that followed immediately after
this read as a bit odd.  Let's revise this to clarify that "Rust" is
the default ABI when `extern fn` is not used.