Convert ABI "details" links to lists
For each of the ABIs, we have outbound links to documents with more
details.  For most of these, we display the URL, since these URLs are
short and readable enough that it's probably more informative to just
show them as compared to trying to describe the document.  However,
the URLs are still long enough that they look better as list items, so
let's switch to doing it that way.