Strengthen wording on "C" ABI
When using the "C" ABI, do we "match" the default ABI chosen by the
dominant C compiler for the target or do we "attempt" to match it?

Tough question.  There are believed to be some exceptions in current
behavior.

However, in the Reference, we generally state things in terms of what
the consensus is on how they should be and then note the bugs, so
let's do that here.  We'll just say that we "match" it.