Revert "Reformat whitespace list for consistency"
This reverts commit 60eb145d7e826d3e8dc8a7051fd5b5e5913070b7.

This re-formats our Whitespace to be centered on Unicode's defintion.

This makes it easy to compare with the standard and helps with
Frontmatter.
Unlike regular Rust, Frontmatter cares about the type of Whitespace.
Even if we want to duplicate the definition, having them formatted
similarly makes them easy to compare.