Split up WHITESPACE
I'm splitting out `HORIZONTAL_WHITESPACE` to make it easier to connect
the concept in frontmatter with `WHITESPACE`.

In doing this, I found it awkward to only pull out part when the
comments are there.
I either needed a redundant comment to start a section, re-arrange out
of order from Unicode, or split out all classes.
I did the latter.