Change underscore to be a keyword
This changes underscore from being a punctuation character to a keyword.
This is intended to help better align with proc-macros, which treat it
as an [`Ident`](https://doc.rust-lang.org/proc_macro/struct.Ident.html).

Note one unusual rule is inline assembly `ParamName` which is
`IDENTIFIER_OR_KEYWORD`. From what I can tell, it does accept `_`, but
the fmt template does not. Templates are not specified in great detail
in the std docs, and don't touch on this fact.

Closes https://github.com/rust-lang/reference/issues/1236
Closes https://github.com/rust-lang/reference/issues/2020