Revise text about capture precision
To the language about a capture path, this branch adds, "..., as well
as any further projections performed by matching against patterns",
and it links to the section on capturing and wildcard patterns.

The link is a bit unobvious, since the section discusses what's not
captured.  The "as well as" phrasing is a bit awkward, and it's
introduced after a comma.  That indicates a non-restrictive clause,
which this is not.

The next rule talks about what place projections are.  It seems that
it might be better to instead extend that language than to try to fit
this in here.  We can say that destructuring can effect a place
projection, and that seems right.  Let's do that.