Revise intro on wildcard patterns and capturing
Let's tighten up the wording a bit in the introduction to wildcard
patterns and capturing to make it better match how we'll present
things below.  We'll remove the "for example..." ahead of the example,
as we're not doing this below, and we'll separately extend the
comments within the examples to include the relevant information.