Add forward reference for `non_exhaustive` exception
The normative rule about destructuring and capturing does not itself
state the exception for `non_exhaustive` enums.  That exception
follows below.  Let's add an admonition to link down to it so it's not
missed.

It's not great having this duplication; it'd be better if we weren't
restating this fact twice.  But it's not clear what to do about this.
The first statement is making a broader statement about tuples,
structs, and single-variant enums.  The section that follows below is
focused on discriminants.  We'll think about this more later.