Revise text about field matching and `..`
Let's clean up the wording of this claim a bit.  Let's put capturing
in the present tense, rather than in the future tense, as that's how
we're going to continue below.  Let's take out, as well, the explicit
lead-in to the example, as we're not generally doing that in these
sections.