Revise `...discriminants.slice-patterns-*` rules
The section here stated:

> Matching against a slice pattern that needs to inspect the length of
> the scrutinee performs a read of the pointer value in order to fetch
> the length. The read will cause the closure to borrow the relevant
> place by `ImmBorrow`.

It then goes on to state exceptions for arrays matched against slice
patterns and slice patterns containing only a rest pattern.

In saying that the pointer value is read, this might suggest that the
pointer is captured.  Actually, though, it's the slice (the pointee)
that is captured.  Let's fix this.

Beyond that, as we saw in an earlier commit, it's better to not lean
on the reader to infer the "relevant place" or when the length needs
to be inspected.  Let's elaborate those details and state the full
rule in one go, upfront, and then state a separate guarantee that
matching an array against a slice pattern does not do a read.

We'll also fix a typo, add rule identifiers, and tighten up the
examples.