Add admonition about the desugaring of destructuring
For our purposes in describing capture paths, we define place
projections.  For the text that follows in the sections below to hold,
we need destructuring to be treated as a place projection for this
purpose.

However, it might seem surprising for destructuring to be in this
list.  While it is believed to have the same effect as a projection
expression, we might not consider an instance of pattern destructuring
to be a projection expression exactly.

To better contextualize this, let's add an admonition that mentions
that pattern destructuring desugars into the kind of field accesses
that we would more likely think about as projection operations.