Clarify struct pattern binding shorthand
The existing text was slightly confusing with the way it worded "ref
and/or mut", and wasn't quite clear what it was referring to. I pulled
that out into a separate sentence. This also adds a sentence to make it
very clear that this is a shorthand.

This is somewhat a duplicate of
patterns.destructure.named-field-shorthand, but it seems fine to me to
have that.

This also reworks the example. The current example didn't fit this
section at all, and it's not entirely clear to me what the original
intention was here. Maybe that was intended for the destructuring
section? Either way, I just made this example fit this rule.