Clarify note about extended temps in initializers
We have a note that describes why, perhaps surprisingly, it's OK to
have a reference to a static with an interior mutable value in the
final value of a constant item but it's not OK if the interior mutable
value is a temporary from the initializer.  There was room to improve
how this was worded, so let's take it.