Add placeholder lifetime to syntax index
The placeholder lifetime is the lifetime equivalent of the inferred
const and the inferred type (and, in fact, we should probably rename
it to the inferred lifetime).  Let's mention it alongside these
similar things in the syntax index.

We could ask whether it should really appear in the "type expressions"
section.  After all, it's not exactly a type expression.  However,
neither is `for<...>`, but that appears in this section.  We can
justify both by observing that they may appear within `impl Trait` and
`dyn Trait` -- and these are type expressions.