Remove "unsafe static items" from syntax index
For the `unsafe` keyword, we had noted that it applied to "unsafe
static items" and linked to the section that describes how mutable
static items must be accessed within an `unsafe` block.

Probably this link doesn't make sense.  The requirement that an
`unsafe` block be used to access mutable statics is just one of the
places where an `unsafe` block is required.

Conversely, if we mean for this to refer to the syntactic fact that
static items can be modified with the `unsafe` keyword, then we'd want
to note this as well for the `safe` keyword.  However, semantically,
we only support these modifiers on external statics, and we already
list those separately.  So let's just remove "unsafe static items"
from the list.