Clarify about `->` and types
For `->`, we had listed "function return type" and "closure return
type".  For me, I read this table as "the thing on the left is used in
the thing on the right".  But that's not really true here; the `->` is
not used *in* the function or closure return type.  It's probably
better just to say that `->` is used in functions and closures, so
let's do that.