Refer to relaxed trait bounds for `?`
For `?`, in this branch, we had listed "questionably sized" in the
syntax index.  I'm not a fan of this terminology.  As I wrote
elsewhere, in Rust PR 145924:

> I don't think we should use the word "maybe" to refer to "?X" bound
> relaxations.  Either there's a predicate stating that a type
> parameter implements some trait or there isn't.  While of course I
> get it -- the type argument provided may or may not implement the
> trait -- I just think "maybe" is speaking to the wrong thing here.
> It focuses on the type argument when what makes more sense to focus
> on is the type parameter.  And the type parameter is not in a
> "maybe" state.

In this regard, I feel the same about "questionably" as I do "maybe".

We don't use this terminology elsewhere in the Reference, so let's not
use it here.  Let's instead refer simply to "relaxed trait bounds", as
we do elsewhere in the syntax index.