Clarify ATBs in syntax index
On the branch, we had an entry in the expressions table for
`Type<ident=Type>`.  This was notated as "explicit associated type
bounds" and referenced a TODO item about default type parameters not
being documented.

Perhaps what this means to describe are associated type bindings and
associated type bounds.  Let's document those in the type expressions
table.  We'll remove the TODO note, as it's not relevant to this.