Add space ahead of curly brackets in index
For macro invocations, struct expressions, and struct patterns, we had
shown the opening curly brace without a space in front of it.  This
can make it visually difficult to distinguish from an opening
parenthesis.  Using the common Rust style of putting a space in front
helps with that, so let's do that here.