Clarify description of items in syntax index
On the branch, in the syntax index, items were described as
"declarations in a crate".  Some items may be declarations, but many
items are actually definitions.  It's probably better to not use
either term here.  Let's instead say that items are the components of
a crate as that's the language that's used in the items chapter.