Go with "constant items" for `_` in index
In the keywords table in the syntax index, every entry except the
one for constant items for `_` consists entirely of a link body.  For
that one, though, we had left "unnamed" out of the link body.
Visually, that just stands out too much.

We could just put "unnamed" in the link body, but most of the other
entries aren't really that specific.  It seems OK to just say
"constant items" for this, so let's do that.