Say "use" rather than "usage" in the syntax index
In each table in the syntax index, we have some bit of syntax on the
left and then on the right we have features in which that bit of
syntax is used.  Should we title the column on the right "use" (or
maybe "uses") or "usage"?

One could almost, superficially, make a case for "usage".  As relevant
here, Webster's defines it as "the way in which a word, phrase,
etc. is used to express a particular idea".  That sounds close.

The trouble is that it's nearly the wrong way around.  The "idea" here
is actually the thing on the right.  The "usage" would be the place or
way in which the syntax is used to express that idea.

Conversely, definitions of "use" apply cleanly: to "employ for or
apply to a given purpose", "an instance or way of using", etc.

As Garner says:

> Whenever *use* is possible, *usage* shouldn't appear.

Let's say "use".