Clarify const item text WRT mutable statics
The rule is that "mutable references contained within a mutable static
may be referenced in the final value of a constant."

However, the key here is that "mutable static" means either a `static
mut` item or a `static` item with an interior mutable type.  We had
made that clear over in `const-eval.const-expr.path-static` but not in
our rules about const item validity, and our link for "mutable static"
had implied it may mean `static mut`.

Let's fix that and add an example to demonstrate the point.