Align `no_std` intro language better and reword
We had said here that the `no_std` attribute is used to "prevent the
automatic linking of the std crate", but then further down we say,
"using `no_std` does not prevent the standard library from being
linked in".

The important distinction here is the word "automatically", but that
gets a bit lost due to reusing the "prevent" phrasing.

Probably "prevent" is just too strong a word to use when what we're
really saying is that we're causing something to not happen
automatically.  Let's make that more clear and increase sentence
parallelism.