Revise `no_std` intro WRT the preludes
We had text in the intro about "deferring to the core crate instead",
but it may not have been sufficiently clear what we mean exactly by
that.  Let's talk specifically about how `no_std` causes us to use the
`core` crate for the standard library prelude and for the `macro_use`
prelude.