Revise `no_std` text WRT `macro_use`
The text here could be read in such a way as to imply that `no_std`
affected whether or not `core` was added to the extern prelude, but
that isn't the case.  Let's make it clear that this section is about
what is added to the `macro_use` prelude and speak to that directly.