Bring `no_implicit_prelude.edition2018` into present
The remainder of this note is in the present tense, so saying that
something "will" happen starting in the 2018 edition makes it sound as
though that will happen in the future, even if what we mean is that
the compiler "will" do it in the 2018 edition and later.  Let's use
the present tense here in emphatic form to distinguish this, and let's
expand "it" for clarity.