Say "dereference expressions" in constant exprs list
To define what's allowed in a constant expression, we have a list that
starts with the text:

> The following expressions are constant expressions...

Correspondingly, most of the items in this list are stated in terms of
being a "this expression" or a "that expression".  However, for deref,
we had instead said "the dereference operator".  Even though the
`expr.deref` section is currently titled "the dereference operator",
it seems more clear and consistent in the context of this list to talk
about dereference expressions, so let's do that.  These are
expressions, after all, The grammar production is named
`DereferenceExpression`.

(There is one other place in the list that we talk about operators:
"built-in negation, arithmetic, logical, comparison, or lazy boolean
operators used on integer and floating point types, `bool`, and
`char`."  Perhaps we'll reword that one later.  It stands out a bit
less than this one did.)