Add note about restrictions on asm attributes
We have a list that describes where attributes may be applied.  We're
adding inline assembly template strings and operands to that list, but
only certain attributes may be used, and other caveats apply.  Let's
note that there are restrictions and link to the relevant section for
more details, as we do already for certain other items on this list.

It may be important, normatively, to note this restriction here if we
now or later refer to some attribute as being accepted anywhere that
attributes are allowed.