Revise `supported-options.nostack`
The phrase "for a function call at the start..." creates a miscue.
One would tend to read "at the start" as restricting the kind of
function call until realizing that doesn't make sense in this context;
then one has to backtrack.  Let's move that clause earlier to avoid
this.

Should some of these be nonrestrictive clauses?  Are they
parenthetical?  We'll say "no" for now, but arguably this could go the
other way.