Revise list item for type alias definitions
In this list, we call many things "definitions" to distinguish that
we're only talking about items here.  While it is more clear to say
"type alias definitions" rather than "type definitions" when referring
to type aliases, let's keep "definitions" in there for now as we use
this terminology in other places, especially related to associated
type definitions, which are spiritually similar.