Remove `derive-helpers` rule with unclear meaning
We're not certain what this rule means; we've run various tests and
could not confirm our interpretation of the rule.  E.g. (thanks to
ehuss for this):

```rust
use pm::{WithHelperAttr, helper};
#[derive(WithHelperAttr)]
#[helper]
#[derive(helper)] // ERROR: expected derive macro, found derive helper attribute `helper`
struct S;
```

Interestingly, this works and calls the attribute macro:

```rust
use pm::{WithHelperAttr, helper};
#[derive(WithHelperAttr, helper)]
#[helper]
struct S;
```

So that we can merge the rest, let's remove this rule for now and add
it back in future work.