operator-expr: remove stray word in footnote
"only for closures that {x} can be {y}" repeats the limitation; it should
either be "only closures that {x} can be {y}", or, if the capability ("{y}") is
implicit, it can be "only for closures that {x}". Go with the first option that
keeps the context, and remove the extra "for".