Fix ambiguity of RESERVED_RAW_IDENTIFIER
This clarifies that the literals expressed in the
RESERVED_RAW_IDENTIFIER rule cannot be followed by a XID_Continue
character. Originally in my mind these literals were to be interpreted
as tokens (and thus assume some kind of break follows them). However,
since this is part of the lexer itself, this doesn't really work for it
to be defined this way.

This helps ensure that strings like `r#_f` or `r#selfie` are not
interpreted as reserved raw identifiers.